Fixed some more linux problems with ptrdiff_t, compiles under linux. Test in Windows pending
git-svn-id: svn://fb09-pasig.umwelt.uni-giessen.de/cmf/trunk@795 a735f7ef-1525-764e-8607-acac38503800

+++ cmf/cmf_core_src/math/integrators/Integrator.cpp
@@ -226,7 +226,7 @@ void cmf::math::Integrator::add_values_to_states(const num_array& operands)
 void cmf::math::Integrator::integrate_until( cmf::math::Time t_max,cmf::math::Time dt/*=Time()*/,bool reset_solver/*=false*/ )
 {
 	m_Iterations=0;
-	ptrdiff_t i=0;
+	int i=0;
 	Time start = m_t;
 	if (reset_solver) reset();
 	if (!dt) dt=m_dt;
